Add custom container and CSS refactor
