+++ src/components/Projects/ProjectsPage.jsx
@@ -1,25 +1,31 @@
+import { Link } from 'react-router-dom';
 import AutoStoriesIcon from '@mui/icons-material/AutoStories';
-import Button from "../reusables/Button/Button";
-import Carousel from "../reusables/Carousel/Carousel";
+import Button from "../../reusables/Button/Button";
+import Carousel from "../../reusables/Carousel/Carousel";
 import styles from "./projects.module.css";
+import Container from '../../reusables/Container/Container';
 
 function ProjectsPage() {
   return (
-    <div id='projects' className={`flex ${styles.projects_container}`}>
+    <Container id='projects' className={`flex ${styles.projects_container}`}>
       <div className={`heading`}>
         <h2 className='flex head'>
           <AutoStoriesIcon />Projects
         </h2>
         <p><b>Recent Works</b></p>
       </div>
-      <div className={`flex ${styles.projects_content}`}>
+      <div className={`flex box_shadow  ${styles.projects_content}`}>
         <div className={`flex ${styles.works}`}>
           <h4>A Record of My Works</h4>
           <p>A list of all my projects, both solo and collaborative works</p>
-          <Button
-            className={styles.btn1}
-            text="See Projects"
-          />
+          <Link
+            to="/projects"
+          >
+            <Button
+              className={styles.btn1}
+              text="See Projects"
+            />
+          </Link>
         </div>
         <div className={`flex ${styles.articles}`}>
           <h4>A Record of My Articles</h4>
@@ -31,7 +37,7 @@ function ProjectsPage() {
         </div>
       </div>
       {/* <Carousel /> */}
-    </div>
+    </Container>
   )
 }
