+++ src/components/Projects/projects.module.css
@@ -1,11 +1,3 @@
-.projects_container {
-  padding: 1rem;
-  background-color: var(--tetiary-color);
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
 h4 {
   font-size: 2rem;
 }
@@ -53,18 +45,8 @@ h4 {
 }
 
 @media screen and (min-width: 768px) {
-  .projects_container {
-    height: 100vh;
-    padding: 1rem;
-    background-color: var(--tetiary-color);
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-
   .projects_content {
-    width: 100%;
-    max-width: 70rem;
+    width: max-content;
     justify-content: center;
     flex-direction: row;
     border-radius: 10px;
