+++ src/components/Skillspage/Frameworks/Frameworks.jsx
@@ -3,7 +3,7 @@ import "../icons.css"
 
 function Frameworks() {
   return (
-    <ul className="flex animate icons-backdrop">
+    <ul className="flex box_shadow animate icons-backdrop">
       <h3>Frameworks</h3>
       <div className="flex icons-container">
         {libraries.map((l) => (
