+++ src/components/Skillspage/SkillsPage.jsx
@@ -4,10 +4,11 @@ import Languages from "./Languages/Languages";
 import Frameworks from "./Frameworks/Frameworks";
 import Tools from "./Tools/Tools";
 import styles from "./skillspage.module.css";
+import Container from '../../reusables/Container/Container';
 
 function SkillsPage() {
   return (
-    <div id="skills" className={`flex ${styles.skillspage}`}>
+    <Container id="skills" className={`flex ${styles.skillspage}`}>
       <div className={`heading`}>
         <h2 className='flex head'>
           <BuildIcon />Skills
@@ -19,7 +20,7 @@ function SkillsPage() {
         <Frameworks />
         <Tools />
       </div>
-    </div>
+    </Container>
   )
 }
