--- src/components/reusables/Button/Button.css
@@ -1,11 +1,0 @@
-.btn {
-  width: 10rem;
-  height: 3rem;
-  font-size: 1rem;
-  font-weight: 700;
-  padding: 0.5em;
-  transition: all 0.4s ease-in-out;
-  border-radius: 5px;
-  background-color: var(--accent-color);
-  color: var(--primary-color);
-}
