--- src/components/reusables/Carousel/carousel.module.css
@@ -1,68 +1,0 @@
-.carousel {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 1rem;
-}
-
-.project_content {
-  background-color: var(--filter-navicons-color);
-  color: var(--cards-color);
-  gap: 0.4rem;
-  flex-direction: column;
-  padding: 0.2rem 0.5rem 0.5rem 0.5rem;
-  width: 100%;
-  border-bottom-left-radius: 1rem;
-  border-bottom-right-radius: 1rem;
-}
-
-.img {
-  width: 100%;
-  border-top-left-radius: 1rem;
-  border-top-right-radius: 1rem;
-}
-
-.project_cards {
-  flex-direction: column;
-  max-width: 25rem;
-}
-
-.project_details {
-  justify-content: space-between;
-}
-
-.tech_grp {
-  gap: 0.5rem;
-}
-
-.arrow {
-  top: 6rem;
-  position: absolute;
-  background-color: var(--tetiary-color);
-  border-radius: 50%;
-  filter: drop-shadow(1px 1px 1px var(--accent-color));
-}
-
-.back {
-  left: 1rem;
-}
-
-.front {
-  right: 1rem;
-}
-
-.slide {
-  opacity: 0.5;
-  transition-duration: 1s;
-}
-
-.active {
-  opacity: 1;
-  transition-duration: 1s;
-  transform: scale(1.08);
-}
-
-.btn {
-  border: 1px solid var(--primary-color);
-}
