--- src/components/reusables/Socialicons/socialIcons.module.css
@@ -1,16 +1,0 @@
-.icon_container {
-  width: max-content;
-  gap: 0.5rem;
-}
-
-.icon_item {
-  align-items: flex-end;
-  justify-content: center;
-  border-radius: 50%;
-  padding: 4px;
-}
-
-.icon_img {
-  width: 1.5rem;
-  filter: var(--filter-icons-color);
-}
