--- src/components/reusables/Socialicons/socialItems.js
@@ -1,40 +1,0 @@
-import twitter from "../../../assets/twitter.svg"
-import github from "../../../assets/github.svg"
-import angelist from "../../../assets/angelist.svg"
-import medium from "../../../assets/medium.svg"
-import linkedIn from "../../../assets/linkedIn.svg"
-
-const socialItems = [
-  {
-    id: 0,
-    icon: twitter,
-    link: 'https://twitter.com/Usirusen_Ikon',
-    name: 'twitter',
-  },
-  {
-    id: 1,
-    icon: github,
-    link: 'https://github.com/UsirusenIkon',
-    name: 'github',
-  },
-  {
-    id: 2,
-    icon: angelist,
-    link: 'https://angel.co/u/usirusen-ikon',
-    name: 'angelist',
-  },
-  {
-    id: 3,
-    icon: medium,
-    link: 'https://medium.com/@usirusenIkon',
-    name: 'medium',
-  },
-  {
-    id: 4,
-    icon: linkedIn,
-    link: 'https://www.linkedin.com/in/usirusen-ikon/',
-    name: 'linkedIn',
-  },
-]
-
-export default socialItems;
